Move BaseDialogFragment to dialogs package
